int.rs: Fix Shl implementation for Int.
DDlog expects the rhs operand of `>>` and `<<` to be of type
`bit<32>`.  Accordingly, we must make sure that all integer
types implement `Shl<u32>` and `Shr<u32>`.  This is what we
did for `Uint`, but not `Int`, which was `Shl/Shr<usize>` instead,
causing Rust compilation errors.

+++ rust/template/differential_datalog/int.rs
@@ -195,21 +195,25 @@ fn test_fromrecord() {
     );
 }
 
-impl Shr<usize> for Int {
+impl Shr<u32> for Int {
     type Output = Int;
 
     #[inline]
-    fn shr(self, rhs: usize) -> Int {
-        Int { x: self.x.shr(rhs) }
+    fn shr(self, rhs: u32) -> Int {
+        Int {
+            x: self.x.shr(rhs as usize),
+        }
     }
 }
 
-impl Shl<usize> for Int {
+impl Shl<u32> for Int {
     type Output = Int;
 
     #[inline]
-    fn shl(self, rhs: usize) -> Int {
-        Int { x: self.x.shl(rhs) }
+    fn shl(self, rhs: u32) -> Int {
+        Int {
+            x: self.x.shl(rhs as usize),
+        }
     }
 }
